Don't copy local image
If the source image is local, instead of copying it over, use qemu-img
command to create an image with the source image as backing image instead.

For crc, this reduces disk usage greatly (~ 9GB currently) and start up
time by around 20 seconds on my very high-end system with an SSD.

Fixes #32.